+++ R/get_trip_geometry.R
@@ -1,27 +1,28 @@
 #' Get trip geometry
 #'
-#' Returns the geometry of each specified \code{trip_id}, based either on the
-#' \code{shapes} or the \code{stop_times} file (or both).
+#' Returns the geometry of each specified `trip_id`, based either on the
+#' `shapes` or the `stop_times` file (or both).
 #'
-#' @param gtfs A GTFS object as created by \code{\link{read_gtfs}}.
-#' @param trip_id A string vector including the \code{trip_id}s to have their
-#'   geometries generated. If \code{NULL} (the default), the function generates
-#'   geometries for every \code{trip_id} in the GTFS.
-#' @param file The file from which geometries should be generated. By default
-#'   uses both \code{shapes} and \code{stop_times}.
+#' @param gtfs A GTFS object.
+#' @param trip_id A character vector including the `trip_id`s to have their
+#'   geometries generated. If `NULL` (the default), the function generates
+#'   geometries for every `trip_id` in the GTFS.
+#' @param file A character vector specifying the file from which geometries
+#'   should be generated (either one of or both `shapes` and `stop_times`). If
+#'   `NULL` (the default), the function attemps to generate the geometries from
+#'   both files, but only raises an error if none of the files exist.
 #' @param crs The CRS of the resulting object, either as an EPSG code or as an
 #'   `crs` object. Defaults to 4326 (WGS 84).
 #'
-#' @return A \code{LINESTRING sf}.
+#' @return A `LINESTRING sf`.
 #'
 #' @section Details:
 #' The geometry generation works differently for the two files. In the case of
-#' \code{shapes}, the shape as described in the text file is converted to an
-#' \code{sf} object. For \code{stop_times} the geometry is the result of linking
-#' subsequent stops along a straight line (stops' coordinates are retrieved from
-#' the \code{stops} file). Thus, the resolution of the geometry when generated
-#' with \code{shapes} tends to be much higher than when created with
-#' \code{stop_times}.
+#' `shapes`, the shape as described in the text file is converted to an `sf`
+#' object. For `stop_times` the geometry is the result of linking subsequent
+#' stops along a straight line (stops' coordinates are retrieved from the
+#' `stops` file). Thus, the resolution of the geometry when generated with
+#' `shapes` tends to be much higher than when created with `stop_times`.
 #'
 #' @examples
 #' data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
@@ -29,6 +30,10 @@
 #' gtfs <- read_gtfs(data_path)
 #'
 #' trip_geometry <- get_trip_geometry(gtfs)
+#' head(trip_geometry)
+#'
+#' # the above is identical to
+#' trip_geometry <- get_trip_geometry(gtfs, file = c("shapes", "stop_times"))
 #' head(trip_geometry)
 #'
 #' trip_ids <- c("CPTM L07-0", "2002-10-0")
@@ -39,19 +44,32 @@
 #' @export
 get_trip_geometry <- function(gtfs,
                               trip_id = NULL,
-                              file = c("shapes", "stop_times"),
+                              file = NULL,
                               crs = 4326) {
 
   checkmate::assert_class(gtfs, "dt_gtfs")
   checkmate::assert_character(trip_id, null.ok = TRUE)
-  checkmate::assert_names(file, subset.of = c("shapes", "stop_times"))
+  checkmate::assert_character(file, null.ok = TRUE)
   checkmate::assert(
     checkmate::check_numeric(crs),
     checkmate::check_class(crs, "crs"),
     combine = "or"
   )
 
+  if (!is.null(file))
+    checkmate::assert_names(file, subset.of = c("shapes", "stop_times"))
+
   # check if required fields and files exist
+
+  if (is.null(file)) {
+    file <- names(gtfs)
+    file <- file[file %chin% c("shapes", "stop_times")]
+
+    if (identical(file, character(0)))
+      stop(
+        "The GTFS object must have either a 'shapes' or a 'stop_times' table."
+      )
+  }
 
   if ("shapes" %in% file) {